modified

+++ service-account/src/main/java/com/gavin/dao/AccountDao.java
@@ -1,6 +1,5 @@
 package com.gavin.dao;
 
-
 import com.gavin.domain.account.Account;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
@@ -8,7 +7,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface AccountDao {
 
-    Long create(Account account);
+    int create(Account account);
 
     int delete(@Param("accountId") Long accountId);
